+++ ExchangeRate.Api/Models/ExchangeRateValueResp.cs
@@ -8,7 +8,7 @@ namespace ExchangeRate.Api.Models
     public class ExchangeRateValueResp
     {
         public string CurrencyCode { get; set; }
-        public decimal initialValue { get; set; }
-        public decimal convertedValue { get; set; }
+        public double initialValue { get; set; }
+        public double convertedValue { get; set; }
     }
 }
